openviking/requirements.txt
openviking>=0.1.12
mcp>=1.26.0
fastmcp>=0.4.0
httpx>=0.27
pyyaml>=6.0
uvicorn>=0.30
